Update patch set 7
Patch Set 7:

> Is there a reason that you can't use getauxval(AT_SYSINFO) instead
 > of "__kernel_vsyscall"?
If I remember correctly，the most troubles are in linker. 
The getauxval implementation depend on the global variable __libc_auxu in libc. but, in the early step of linker init, this variable is not initialized. 

Certainly,  I think it is easy to fix this issue, just initial it at the entry of userspace before init main thread. Or, provide a new global variable to save the KernelKernelArgumentBlock parameter instead of save it in main thread well-known TLS slot, we can call the method getauxval of KernelArgumentBlock to get the entry address in early time.

extern "C" unsigned long int getauxval(unsigned long int type) {
  for (ElfW(auxv_t)* v = __libc_auxv; v->a_type != AT_NULL; ++v) {
    if (v->a_type == type) {
      return v->a_un.a_val;
    }
  }
  errno = ENOENT;
  return 0;
} 
or
class KernelArgumentBlock {
  .....
  unsigned long getauxval(unsigned long type, bool* found_match = NULL);
  ....
}

Patch-set: 7
Label: Verified=0