Update patch set 1
Patch Set 1:

Fwiw, the motivation to get down to one dlmalloc implementation was to guarantee the introspection APIs were going to work if we say walked the native/bionic dlmalloc heap and the ART one using the same inspect_all call. Ideally the code would live in one place to keep this property.
The pixelflinger code is only used in the emulator and during device bring up. Its hard to eliminate the use of dlmalloc as you need the memory to be allocated executable. That said, this issue isn't mission critical.

Patch-set: 1