Update patch set 1
Patch Set 1:

I don't really like this method. Should I create a library of nofortify tests to do this better? The only reason I didn't do that is that fortify is set on the command line for the host tests, and I was afraid we might do that other places which would enable fortify behind our back.

Patch-set: 1