Update patch set 1
Patch Set 1:

> I don't really like this method. Should I create a library of
 > nofortify tests to do this better? The only reason I didn't do that
 > is that fortify is set on the command line for the host tests, and
 > I was afraid we might do that other places which would enable
 > fortify behind our back.

dunno. the nice thing about the big hammer i mentioned on the bug is that we never have to think about it again. if we go this way instead, we'll have to remember whenever we add a new fortified function (if it's not in one of the existing headers, which is one thing that makes what you've done a lot better than what i'd imagined this approach would look like --- the list of headers is smaller and more static than the list of functions).

i guess an advantage of this is that we don't need to change CTS, and we'll regularly run both sets. (and so fortify/non-fortify differences will be pretty obvious.)

can you explain why you don't like this? because it's better than i'd imagined (i was already resigned to the big hammer being the only choice), i kind of like it. i guess you had higher expectations?

Patch-set: 1