Update patch set 1
Patch Set 1:

Mostly I'm not a fan of the including of the source from another file. Also, the compiler might choose not to use the fortify version and you don't have that much control.

I felt like this was a good compromise to get something done quickly and we can think about a better solution later.

I forgot to mention that I tested this on a fugu, put back the bad strrchr, and verified that the regular test passes, but the no_fortify version fails.

Patch-set: 1