Update patch set 1
Patch Set 1:

Does this cause any noticeable slowdown? I don't think we have that much output so it's probably fine, but I could imagine this might run really slowly on something like a watch.

I don't know if there is a faster way to post-process, but if it's really slow maybe there is a way to pre-process.

Patch-set: 1