Update patch set 1
Patch Set 1:

There is already a work around for this in some for the ART buildbot testing. This patch provides a better fix, by using functionality provided by ld.so as a reference.

As you are probably aware, the builbot configuration for the ART testing uses a custom ANDROID_ROOT. On 32bit platforms it all works fine. However on 64bit platforms, when dalvikvm64 (64bit ELF) requests dex2oat (32bit ELF) to build a particular OAT file it all fails.

This has to do with Android's linker failing when in loads the first library that matches the required name.

$export LD_LIBRARY_PATH=$ROOT/lib64:$ROOT/lib
$dalvikvm64 -cp example.dex example
# dalvikvm64 (64bit ELF) starts and it loads fine the libart.so (64bit ELF) $ROOT/lib64/libart.so
# now dalvikvm64 realises that there is not oat file for the example.dex and it invokes dex2oat
# dex2oat (32bit ELF) starts and tries to load libart.so ****KABOOM**** - libart.so is a 64bit ELF file ($ROOT/lib64/libart.so)

ld.so would carry on searching and finally pick up the 32bit ELF libart.so ($ROOT/lib/libart.so). 

The work around for the ART testing would build dex2oat as a 64bit binary. This way when dalvikvm64 execs dex2oat the required libart.so would be in both cases 64bit.

I hope this explanation helps. Kevin will be looking at adding a test that demonstrates this problem.

Patch-set: 1
Label: Code-Review=0