Update patch set 1
Patch Set 1:

> findfp.c seems like an odd place to put fclose/fileno

my plan is to kill the pool at some point. so having fclose with the allocating part of fopen is actually part of a plan.

fileno less so.

i've been trying to decide for a long time whether we go for "stdio.cpp" or not. (i'm assuming we won't rewrite the printf/scanf family, in which case the rest of stdio is all pretty trivial.) i'll upload another change that does that for now, and we can worry about doing something more fine-grained if it ever gets to the point where we need it.

(the stdio/ directory is a historical accident. originally all of bionic looked like this. when i started the whole upstream-* thing, i'd leave bits behind like this until i could switch to the upstream versions. these never made it, and now i feel like there are some parts of stdio we'll just own.)

Patch-set: 1