Update patch set 5
Patch Set 5:

(4 comments)

As I am running the tests using a custom ANDROID_ROOT, most tests failed for me for other reasons so I did not see these failures, sorry for that. I fixed the issue (hardcoded /system/bin/sh in libc), I'll push a patch for this later.

I fixed all the tests, most of them were failing only because they expected dlerror() contents that the patch changed. Matching a substring sounds more reasonable to avoid further failures.

For dlext.ns_shared and dlext.ns_isolated, one of the test was failing because of a slight change of behavior when the caller has its DT_RUNPATH set (in this case libnstest_root_not_isolated.so), there used to be a specific test in open_library() to check whether the file is accessible or not before trying to load it, but since the check is done when loading it anyway I removed this check. For this reason, the dlopen() at dlext_test.cpp:892 now finds a file and fails to load it, instead of pretending it hasn't found any.

Along the way, I also fixed some tests in dlfcn_test.cpp that did not take ANDROID_ROOT into account.

Patch-set: 5