Update patch set 11
Patch Set 11:

> > > > Uploaded patch set 10.
 > > >
 > > > If deprecate int80 completely, [...]
 > >
 > > If by "deprecate int80 completely" you mean "require AT_SYSINFO
 > to
 > > function", please don't.  It's possible to configure a kernel at
 > > runtime or boot time to have no vdso, and IMO a libc
 > implementation
 > > should still work.
 > 
 > but why would Android want to support that? we already have a whole
 > list of things you need to have in your kernel. this one seems
 > worth adding to the list.

If you want to run Android and OpenSUSE 7.3 or whatever it was, you'll need it :)

More seriously, the ability to boot with vdso32=0 is occasionally useful for testing.  And isn't supporting this about one line of code plus an asm helper that is three bytes long?

Patch-set: 11
Reviewer: Andrew Lutomirski <1090553@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0