Update patch set 1
Patch Set 1:

(9 comments)

i think last time these came up, we decided not to add them on the assumption that they'd be removed in the next standard. and because they're pretty much useless (if not dangerous) on Android because all app code is running in a multi-threaded context.

but i guess the __deprecated__ might be sufficient.

what does everyone else think?

Patch-set: 1