Update patch set 5
Patch Set 5:

The background here is that LTP uses this API (not just to test it, but internally in other tests and in libltp).

I considered just open-coding private versions of these in the LTP tree.  But since these are still required parts of POSIX, if I'm going to go to the trouble of implementing and testing them, it seems like a waste not to take them into bionic.

Patch-set: 5
Reviewer: Greg Hackmann <1015340@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0