Update patch set 1
Patch Set 1:

The issue we are seeing with mips now is related only to mips64, 
so if it would help, we could enable clang for mips32
(after shines cherry-picks necessary clang changes).
Not sure if it is worth the effort, though.

Patch-set: 1
Reviewer: Nikola Veljkovic <1062659@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0