Update patch set 3
Patch Set 3:

> Couple things:
 > #1: not sure which architectures have builtin nearbyint routines. 
 > Is there a way to know without compiling and checking the objdump?

experience with fabs showed that it's not so much "does the compiler have the support?", it's "is the compiler's implementation any good?". so you'll need objdump regardless.

that's why this work is slow :-(

 > #2: Do you want me to only refactor these to cpp or do you want me
 > to do all of the existing ones as well?

if you want to do the existing stuff, be my guest. otherwise we'll get round to it at some point. but for _new_ stuff i think clang is the right place to make sure we have the best intrinsics, and bionic should just get out of the libm optimization business.

the current situation hurts everyone really, because most code _isn't_ calling libm --- it's getting compiler intrinsics inlined. so if they suck, all realistic code sucks no matter how good libm is. i should perhaps have taken a stronger "look, clang's arm fabs intrinsic is broken, please fix it" stance. it felt at the time like i was being the good guy by having the special case in libm, but the more i've thought about it, the more i think it's a mistake because it means our microbenchmarks give us a warm fuzzy feeling that doesn't reflect the performance of actual compiled code.

Patch-set: 3