Update patch set 2
Patch Set 2:

> (3 comments)
 > 
 > the backwards naming still seems a bit weird. should we always
 > output oem1_xxx and oem2_xxx but interpret legacy oem_xxx as
 > oem2_xxx until brillo can move over?
 > 
 > what do the brillo folks think? how many dependencies on the
 > current naming scheme (where the second range is oem_xxx) do you
 > have?

We have no dependency on any specific mapping (we don't care what the final UIDs are), but we do have a dependency on being able to access a range of 1000 UIDs using string names like "oem_XXX". We could convert our stuff to use "oem{1,2}_XXX", so if you want to change oem_XX to be the first range, and oem2_XXX to be the second range, now's a good time to do it. We currently have no hard-coded "oem_XXX" users, but will soon-ish.

Patch-set: 2