Update patch set 3
Patch Set 3:

Crossing this bridge too soon.

Do not worry about logic, if logic can make things more maintainable, I am all for it.

We need something better for the immutable set.

[TL;DR]

The fs_config location is a nice touch, but I find a C file (even the existing one) easier to maintain than an ini format file. The AIDs are listed in the same file as the names in the original file, one place. Having an ini file in another place, another repository, from the AID definitions, then used to generate the even uglier pair of passwd/group files full of difficult to parse immutable content for runtime consumption feels like a step in the wrong direction.

The rules are AID number and name, that is all we really need, everything else is chipped in stone, the clutter of root directory, shell and other passwd fields gets unruly so a passwd _and_ group file is also equally annoying and just adds time to the linear search for the other ids. I know we can search android_ids[] in 25ns for a match. Going out to a file costs us 1.1us, even higher if the file is allowed to grow as you are suggesting. I am not extolling complexity, just pointing out the cost of trying to centralize on /system/etc/passwd.

I do admit I envisioned it being tiny and almost directly controlled by partners, and limited in scope. Limited to 1099 overrides and that is it. No other options to the partners for any of the other AIDs. I do want to cut down on duplication though. However, names are generally only picked up to pretty up some references to human readable values so the cost could be considered meaningless.

libpackageparser is even more expensive for the dynamic application ids; fortunately we will not go out on this limb except on libc+libpackageparser if we follow the plan; the cost is controlled.

We will always have the logic in grp_pwd.cpp, having ranges for each Android grouping of AIDs to generate the default names. android_ids[] sits in that nether-land of default, immutable and ubiquitous in Android.

Patch-set: 3