Update patch set 3
Patch Set 3:

> Here is what I envisaged because of architectural constraints and
 > concerns, you lit the fire under the deprecation of
 > android_filesystem_config.h android_ids[] ...
 > 
 > In android_filesystem_config for the most part, android_ids[] can
 > be intuited by AID_<UPPERCASE> -> "<lowercase>" except for:
 > 
 > AID_MEDIA_DRM -> "mediadrm"
 > AID_MEDIA_EX -> "mediaex"
 > AID_MEDIA_CODEC -> "mediacodec"
 > 
 > I see no problem with "media_drm", "media_ex" and "media_codec"
 > respectively moving forward so that the rules are set. One merely
 > adds new immutable Android AIDs to android_filesystem_config.h, and
 > no longer worries about filling in android_ids[].

Trivial to collapse _ if we want, but that's superfluous at the moment (easy change).
 > 

Let me see what types of support Python has for parsing header files, I wanted to avoid
that as much as possible. But auto-gen an AID_* header file, not sure how to make the deps
work properly for #include users...

 > Consider a generator to take that and construct an android_ids[]
 > (like) creature for grp_pwd_cpp _only_ to consume. Chipped in
 > stone, landing in binary, efficiently searched, as part of libc.so.

Nods.

 > 
 > From there passwd/group get only AID_OEM_RESERVED_START to
 > AID_OEM_RESERVED_END, and AID_OEM_RESERVED_2_START to
 > AID_OEM_RESERVED_2_END, from the base ini files already constructed
 > to support fs_config. fs_config will only allow this set to be
 > added, and places strict rules on the name length (e.g. 32
 > characters).

To be clear, OEMs still use the ini format, and we consider them when producing
the static android_aids array?

 > 
 > If you feel like writing that tool that translated
 > android_filesystem_config.h into a built copy of android_ids[] for
 > local grp_pwd.cpp construction; we eliminate the fs_config
 > 'defaults' ini file and centralize on include/private/android_filesystem_config.h
 > (with a note indicating that things other than C will parse that
 > include file).

I will do that, I just need to see whats the best way. I don't want to pull in any
external python package for this. AOSP does not need yet another dep. Granted
I could write the parser, and like you said, comment that header file. I like
this as it avoids the complexity of dealing with auto-generated header file
dependencies, but I would at least like to consider taking the base config ini
and generating the header file, but the Make system im not sure of. We could
just #include <generated_aids.h> into android_filesystem_config.h but how to handle
deps? I would at least like to hear comments (please) on this, juts for my
understanding.

 > 
 > Others may have problems with adding logical complexity ... so this
 > is just my opinion.

I'm fairly ok with this, it doesn't allow for AID_FOO for oems to be magically
there via android_filesystem_config.h, but that's a nice-to-have IMO.

Patch-set: 3
Label: Code-Review=0