Update patch set 6
Patch Set 6:

The lp32 version is not exposed in NDK (starting with r11 at least).

The question is do we want to expose it in NDK for version prior to O? If yes - move this function to libc section for lp32 while keeping libc_o version for lp64.

If not - we should keep it as is.

I am leaning toward add everything to LIBC_O, after all it was not functional before.

Patch-set: 6