Update patch set 5
Patch Set 5:

(1 comment)

See inline comment. I'm not sure that flushing the cache is necessary. Even if it is correct to do so in some cases, it doesn't seem worth the extra complexity (or the performance impact).

I would just unconditionally update the search path every time this function gets called and not change the logic for when we decide to flush the cache.

Patch-set: 5
Label: Verified=0