Update patch set 7
Patch Set 7:

Well, it is also a matter of control I guess, drawing a line in the sand. jemalloc is the _only_ external (from bionic repository) static library incorporated into bionic libc so it felt like a precedent to include it. libpackageparser is one source file with two simple functions; not insurmountable to incorporate and when all is said and done, will end up being internal+private.

The goal is to wean all native functions off of special Android interfaces and use the POSIX getpw*/getgr* functions. If libpackageparser continues as a _public_ interface, we have failed.

Last two patches in this current series are still a (refined) prototype for now.

Patch-set: 7