Update patch set 7
Patch Set 7:

The unused fields (from the grp_pwd) perspective are debuggable, seinfo, gids and private_data. These fields are however used by run-as and libselinux; so we let them be collected and not used in this context?

I believe enh@ is making the point to strip the internal (and hidden symbols too) bionic packagelist parser to a bare minimum, and let the system/core/libpackagelistparser be used by libselinux and run-as. If that is the case, I do not like having to maintain/support two.

I set the precedent to use private libc interface in liblog (log_is_loggable utilizes private properties), and libcutils uses them as well (to provide the well known APIs to properties). I see no problem with using the private libc libpackageparser interfaces in libselinux and run-as with the goal of removing system/core/libpackagelistparser.

Patch-set: 7