Update patch set 6
Patch Set 6:

This has a conflict on the move to Android.bp files. Im not sure how to tie in the file generation. I see that you can specify deps and add commands via Ninja (https://ninja-build.org/manual.html). But looking at the blueprint files, I can see that its not a straight usage of Ninja, but rather theirs a host of generators written in Go that are plugged into it (things like cc_library_static).

If someone could point me to an example on how to re-write the generated sources into the new blueprint format, it would be greatly appreciated.

Patch-set: 6