Update patch set 6
Patch Set 6:

Generated files are still a work-in-progress, but if they become a blocker for this I will get them fixed.

As a side issue, one of the goals for Soong is to have more deterministic builds, including across devices - ideally libc.so on different devices with the same CPU will be identical.  Compiling arbitrary device-specific data into bionic is going to prevent that, so I'd prefer to see an external data file, but that's really up to Elliott.

Patch-set: 6
Label: Verified=0