Update patch set 1
Patch Set 1: Verified+1 Autosubmit+1

I'm open to other suggestions on this one. We could do the same thing in the build system, but that isn't much better. I suppose it means we won't accidentally ship this in the NDK, but it also means we'd have to keep it in two places (make and soong).

Patch-set: 1
Reviewer: Dan Albert <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=+1
Label: Verified=+1