Update patch set 1
Patch Set 1:

> > sweet. is that already in upstream, or something that's being
 > > worked on?
 > https://reviews.llvm.org/D14274 -- the LGTM is a lie (Richard still
 > needs to LGTM it). The patch also exposes some ugly behavior in
 > Clang's CodeGen though, so we need to fix that before committing
 > it. Hopefully I'll get to it soon. :)
 > 
 > > (i.e. do i have a revision i should be watching for to
 > > come back and fix this?) or is there a __has_feature test?
 > 
 > #if __has_attribute(alloc_size)
 > ...
 > #endif
 > 
 > :)
 > 
 > If you want, I can make that part of the FORTIFY changes I'm doing,
 > which should land after we pull in a clang version from after June.

i was actually going to do the opposite: add a timebomb to force me to remove the #if when we have a clang that supports this :-)

but, yeah, if you've already got a change that requires the future clang, adding this there does sound like the best way to track this. thanks!

Patch-set: 1