Update patch set 2
Patch Set 2:

cc_benchmark should be building for both, I have that change in a larger patchset but it might take a while to land.  Feel free to upload a fix to cc_benchmark if you want it in sooner.

Patch-set: 2
Reviewer: Colin Cross <1002751@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Verified=0