Update patch set 2
Patch Set 2:

> I added a abort() at the end of TEST(UNISTD_TEST, execvpe_ENOEXEC),
 > and I see the test fails because of abort, and strace -f reports
 > line:
 > 
 > [pid  7012] tgkill(7012, 7012, SIGABRT) = 0

but we're not using ASSERT_EXIT any more. it works like this, but the ENOEXEC test wasn't running anything other than the first success case when i wrote it but kept the ASSERT_EXIT implementation. (which is really scary because you couldn't even tell. it might be worth factoring out the setup so i can split this test into multiple tests just in case.)

Patch-set: 2