Update patch set 1
Patch Set 1:

Thanks to Stephen and Pirama.
The bug fix for this already has been applied to the truck of llvm (r277501), hence the bug report (https://llvm.org/bugs/show_bug.cgi?id=30308) is now marked as solved. This commit can be reverted or stayed at master as it is since NFC.

Patch-set: 1