Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+2
> 
> aren't we better off just doing this the simple static_assert way like the pthread types? knowing the actual sizes doesn't really help you debug a problem anyway.

Well, it was certainly nice for knowing what the size/alignment was so I could set up the struct in the first place. If it were to somehow change in the future I suppose it should be immediately obvious from the diff, so I guess we don't need to keep it this way.

> (maybe rename to <bits/struct_file.h> too?)

Can do.

Patch-set: 1
Label: Verified=0