Update patch set 1
Patch Set 1:

> So, there's still a shit load of tests that fail on the host even
 > after this. Those used to all be green (I haven't run them in over
 > a year at this point). Did we give up on that?

they've been green now and then, but never consistently because it's not part of anyone's regular workflow.

i'm assuming we'll fix that as part of dwillemsen's work, and i also assume that as part of that we'll build icu4c for the host too.

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0