Update patch set 9
Patch Set 9:

No, a different one. Fixing and dealing with this is a low priority, getting the first one of this patch series in before addressing selinux adjustments. I did add a mitigation in this version of the code to stop after being blocked the first time so the violations are rare. I expect the _real_ fix will be to detect permissions or context first, and without inciting an actual selinux access violation, rather than a dontaudit.

Patch-set: 9