Update patch set 1
Patch Set 1:

Would it be possible to have some kind of verification of the internal thread before use? It's possible that a use after free might not crash, but instead use some new garbage data.

Perhaps verify that the tid or cached_pid are correct (and clear those values on free).

Patch-set: 1