Update patch set 1
Patch Set 1:

Yeah, you probably don't need to have the free set any variables, but I still worry that a use after free might kind of work. It might not cause a crash if something snuck in and mmap'd something new in the place of the empty map.

As an app developer, I worry because it would be really difficult to debug that kind of problem.

It would be interesting to know if destroying and creating threads ever winds up recreating the same map as before. If so, would it be more likely that someone attempting to use a freed thread_t doesn't get a crash, but goes off the rails in some mysterious way.

Patch-set: 1