Update patch set 1
Patch Set 1:

> It would be interesting to know if destroying and creating threads
 > ever winds up recreating the same map as before. If so, would it be
 > more likely that someone attempting to use a freed thread_t doesn't
 > get a crash, but goes off the rails in some mysterious way.

Wouldn't the behavior if this happened be the same as it would before? i.e. operating on the new thread identically as if it were the old thread

One alternative would be to store a checksum or magic number in the pthread_internal struct and check that wherever we used to look up the thread in the list, but I don't know if that's worth doing.

Patch-set: 1