Update patch set 2
Patch Set 2:

ps2 just updates the checkin comment to explain more of the thinking.

i wrote a test to answer cferris' address-reuse question, and the answer is "yes, a new thread gets the previous thread's mapping (both LP32 and LP64)". so the problem he was worried about (unsafe code accidentally manipulating the _next_ thread) is a problem with or without this change: the new thread will have got the same mapping and been added to the global thread list, so we've no way of telling that it's not the same thread.

+a few others, since this existing problem is not (i think) well known, and is worse than the tid reuse problem (since the kernel will cycle through the range before reusing).

Patch-set: 2
Label: Presubmit-Ready=0