Update patch set 2
Patch Set 2:

pthread_t is always allocated at the top of the stack. Maybe we can add a random shift when allocating pthread_t (like from 0-1k bytes). And we can add a magic number in pthread_internal_t as josh said. So if someone is trying to use a deleted pthread_t, either it crashes (because the space is unmapped), or we know it is not valid.

Patch-set: 2
Label: Verified=0