Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> __attribute__((packed)) doesn't imply an alignment of 1 for the whole struct, right?

It indeed does imply an alignment of 1. Both Lorenzo and Elliott (on a system/netd CL) said that these structures are internal only, and only accessed in a way that did not require the packed attribute. Do you think that a 1-byte alignment is necessary for either of these data structures?

Patch-set: 1