Update patch set 2
Patch Set 2:

Note that this is a tad premature: at the very least, https://android-review.googlesource.com/#/c/335499/ still needs to go in.

Also, the test line is a lie: the builds/tests are running on my desktop as I type this. Depending on how those go, we might still have a few patches remaining, but we should be just about there. Finally. :)

Patch-set: 2
Label: Verified=0