Update patch set 1
Patch Set 1:

> There are a few places where g_argv[0] is used in log messages, is
 > that going to be a problem? I assume that %s of nullptr is going to
 > print (null), but I don't know if every log will handle that
 > correctly.

yeah, we'll do the "right" thing there.

the alternative would be to do what __libc_init does, which is to use "<unknown>" instead if it's given nullptr (you can see that in the run-as output). but that's not obviously better (and "(null)" is actually a better clue as to where you went wrong).

Patch-set: 1
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>