Update patch set 2
Patch Set 2:

I believe the failure is because this is merging to master but is part of a topic. Since the other files are needed to let this compile correctly, treehugger is not happy.

I'm going to verify this myself by cherry-picking to internal and rebuild.

Patch-set: 2
Reviewer: Elliott Hughes <1003224@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0