Update patch set 5
Patch Set 5:

> here's the complaint:
 > 
 > frameworks/native/libs/vr/libbufferhubqueue/buffer_hub_queue_client.cpp:52:34:
 > error: constant expression evaluates to -2147483647 which cannot be
 > narrowed to type 'uint32_t' (aka 'unsigned int') [-Wc++11-narrowing]
 > epoll_event event = {.events = EPOLLIN | EPOLLET,
 > ^~~~~~~~~~~~~~~~~
 > bionic/libc/include/sys/epoll.h:43:26: note: expanded from macro
 > 'EPOLLIN'
 > #define EPOLLIN          0x00000001
 > ^
 > 
 > the weird part being that EPOLLIN didn't change. but EPOLLET
 > changed to (1<<31) instead of 0xblah. but i don't understand why
 > that should make any difference from clang's point of view? aren't
 > both ways of writing it implicitly signed?

Sorry, I was out last week. The problem is that 0xblah is acceptable as an unsigned constant, but (1<<31) is going to be treated as a signed value (with undefined behavior on top of everything else, since you shifted a known different value into the sign bit). Thus, the change to (1<<31) does make the result signed, which then is inappropriately narrowed to the unsigned result. This is all a bit unfortunate, but the proper fix should be to continue using 0xblah for constants like this.

Patch-set: 5