Update patch set 5
Patch Set 5:

Oops, sorry, I meant the this might be good for a cortex-a57 specific routine.

But as I mentioned, this is not good for a big-little (a53-a57) device since it is a large enough regression on the little core that it doesn't make up for the performance improvement on the big core.

Patch-set: 5
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>