Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Are there any A57 devices without A53?  Surprised that this is a regression on Marlin if it's good for big cores (Kryo is closest to A57 as far as I know).

If the a57 speed-up is large enough, and you wind up on the big cores enough, you probably wouldn't notice the regression due to the little core regression. Unfortunately, the bionic benchmarks code is not ideal for doing benchmarking on asymmetric cores. I hope to have time to fix up the bionic benchmark better to handle this in the future.

Patch-set: 5
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>