Update patch set 5
Patch Set 5:

(1 comment)

This routine is slightly worse for cortex-a57 cores, not terrible, but enough that it's noticeable.

Also, there is a small regression (~3%) for the cortex-a53 core, for sizes >= 16 bytes to about 1024 bytes. Not all values in this range, but enough that this makes me nervous because I don't have a good feel for the calls to memset and the distribution of these call sizes.  There is a relatively big win for the larger sizes though (~10%, not quite as big as measured on the Nextbit).

I'll do further classification to see if I can figure out what's going on.

Patch-set: 5
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>