Update patch set 7
Patch Set 7:

The only reason I did it this way (NEON and non-NEON) is so that it can automatically be used for all ARM targets instead of having to maintain the target-specific lists.  The ARM emulator doesn't assume VFP+NEON, so build would fail.  I can just do the NEON version and add to all the subarch lists though if that's preferred.

Patch-set: 7
Reviewer: Jake Weinstein <1021512@85c56323-6fa9-3386-8a01-6480fb634889>