Update patch set 1
Patch Set 1:

Just some tests to make sure the functions are linked in and available.

Elliott usually just adds some calls that return errors, to make sure that we didn't forget to export them, and that they don't get broken in some weird way in the future.

If a function can have some kind of smoke test that would be nice, but not absolutely necessary. I think we've done an EINVAL check and done something slightly different if the function is supported.

Patch-set: 1
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>