Update patch set 21
Patch Set 21:

By the way, I think the memmove comparison is invalid. It uses memcpy so, comparing a before/after would show improvements based on a different memcpy, not on the memmove change. In fact, I'm not sure that the bionic benchmarks even exercise the changed code.

The micro_bench tool provides a memmove_backwards and a memmove_forward so you can compare performance of the function operating with overlapping and non-overlapping buffers. As I mentioned previously, at some point this summer, I hope to enhance the bionic benchmarks so that it will give you all of the information and we can get rid of micro_bench.

Patch-set: 21
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>