Update patch set 4
Patch Set 4:

> Never mind, that's the wrong repo.  Where did qemu move after
 > external/qemu was deprecated?

+bohu?

was it deprecated, or is it just no longer part of the master manifest because master uses prebuilts? i think it's the latter.

i filed http://b/18843629 in 2015 asking the emulator folks to unfork, but it doesn't look like anything's happened yet.

bohu: does external/qemu's copy of libjpeg build for the host or the target?

Patch-set: 4