Update patch set 17
Patch Set 17: Code-Review+1

> > It should be noted that we have no passionate consumers of this
 > > enhancement to grp_pwd, thus making a case for this difficult to
 > > rationalize.
 > 
 > Intel is a passionate user. The other set of merged patches someone
 > from Sony ack'd on it.
 > 
 > I'd like to have this in-my-tree but as a general rule, I don't
 > cherry-pick anything that isn't in AOSP.

switching to "oem_12345" style user names in scripts is not really an option as we have many users and scripts. without this change we'd have to continue to patch in usernames in system/core (or into the script in build) which would be good to not have to do.

Patch-set: 17
Reviewer: Johan Redestig <1002951@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1