Update patch set 5
Patch Set 5: Code-Review+2

We can always retcon this into the NDK, but that doesn't need to happen right now.

To do that, you'd add `# versioned=27 introduced=26` to the symbols. You should test that with readelf to make sure I'm not remembering wrong, but that should work (if it doesn't I should fix the build system so it does)

Patch-set: 5
Reviewer: Dan Albert <1043845@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2