Update patch set 2
Patch Set 2:

> The problem is that this code is not used by any currently tested
 > config. So even if you make this change, we don't compile it by
 > default, and we don't even know if it works.

This change can be tested with the modified aosp_arm64 (also in a CL of this topic). How do we add it to the tested configs (if not already)?

 > At some point, we'd like to get rid of this file completely.

Then what strlen implementation is used for ARM CPUs other than the ones listed in current Android.bp?

Patch-set: 2
Reviewer: Isaac Chen <1139803@85c56323-6fa9-3386-8a01-6480fb634889>